--- a/src/core/MOM.F90
+++ b/src/core/MOM.F90
@@ -2430,12 +2430,12 @@ subroutine initialize_MOM(Time, Time_init, param_file, dirs, CS, restart_CSp, &
 #endif
   G_in => CS%G_in
 #ifdef STATIC_MEMORY_
-  call MOM_domains_init(G_in%domain, param_file, symmetric=symmetric, &
+  call MOM_domains_init(G_in%domain, US, param_file, symmetric=symmetric, &
             static_memory=.true., NIHALO=NIHALO_, NJHALO=NJHALO_, &
             NIGLOBAL=NIGLOBAL_, NJGLOBAL=NJGLOBAL_, NIPROC=NIPROC_, &
             NJPROC=NJPROC_)
 #else
-  call MOM_domains_init(G_in%domain, param_file, symmetric=symmetric, &
+  call MOM_domains_init(G_in%domain, US, param_file, symmetric=symmetric, &
                         domain_name="MOM_in")
 #endif