### This PR introduces two enhancements:

 1. Automatic Mask Table Generation: This feature allows for the elimination of land blocks by automatically generating a mask_table during domain initialization at runtime. It eliminates the need for any preprocessing steps and tools, simplifying the user experience. It also ensures that the initial PE count set by the user is fully utilized, thereby eliminating the need for a clean re-build in CESM. Users can activate this option by setting the `AUTO_MASKTABLE` parameter to True. Relevant commits:
   - [first working version of an automated mask table generator](https://github.com/NCAR/MOM6/commit/f2f4f0414af02e98578e938084c11aa8dbdfa203)  
   - [While determining masked blocks, take reentrancy and tripolar stitch …](https://github.com/NCAR/MOM6/commit/4202b78e8ef61577c3ebcb7b3a165964d6c13b9b) 
   - [apply tripolar stitch fix in auto mask_table generation](https://github.com/NCAR/MOM6/commit/3cb66ec9cc9a4c244123cf3163fe19916e807667)
   
 2. Land block elimination support in the NUOPC cap. Relevant commits:
   - [Enhancements for adding land block elimination to NUOPC cap:](https://github.com/NCAR/MOM6/commit/c58aa8ddc3b298685ae9c1daa08abd9badcc825b)
   - [Enhance NUOPC cap to support MOM_mask_table](https://github.com/NCAR/MOM6/commit/b017639c9c2a2b4ee94ca134c602c12f317ded8e)

### More on how automated land block elimination works:

- Based on the target PE count `npes` (set by the user), calculate an upper limit on the potential number of new domain divisions: `npes/glob_ocn_frac`, where glob_ocn_frac is the ratio of total ocean cells to total cells. While this upper limit may be overly optimistic for realistic domains, it can be achievable for idealistic domains.
- Starting from this upper limit, iteratively explore different domain division counts `p` to determine the layout and identify the maximum number of blocks that can be eliminated to meet the target `npes`. Once this condition is met, finalize the iteration and adopt the determined `p` as the new number of divisions (where `p` - `npes` correspond to the number of masked blocks.)
- For FMS to read back in, write out the mask table based on the number of divisions determined in the previous step and adopt the corresponding layout as the new domain layout.

This entire iteration is quite fast, taking less than 0.1 seconds for our 0.66-degree workhorse grid.

### Performance results:

When auto land block elimination is turned on, **we get 20 to 23% speed up.** Below table summarizes the model throughput (simulated years per day) for 3-month long CMOM_JRA.TL319_t232 runs on derecho.intel with various target PE numbers.

|  MOM6 PEs | throughput (base) | throughput **(auto LBE on)** |
| - | - | - |
| 640 | 18.89 | 24.29 |
| 896 | 25.75 | 32.11 |
| 1152 | 31.18 |  40.01 |
| 1280 | 34.62 |  42.15 |
| 1408 | 37.19 |  46.81 |

- Below images display the land block elimination for two select PE counts, 896 (default) and 1408 (highest tried):

896:
![pe896](https://github.com/NCAR/MOM6/assets/12533965/3c001cfa-a06e-4313-92eb-6e77adca71d6)

1408:
![pe1408](https://github.com/NCAR/MOM6/assets/12533965/d099eaa5-d653-4c3a-abeb-942361b5ead2)



 

 
### Potential to-do items:

 - We may consider moving the newly added subroutines (for auto masking) from the `MOM _domains` module to a new module, say `MOM_auto_mask_table`. Note that this would necessitate moving the `MOM_define_layout` subroutine to somewhere else (`MOM_domains_infra`?) to prevent circular module dependency.
 - We may consider turning some hard-coded parameters and filenames (e.g., `ibuf`, `r_extreme`, and  `auto_mask_table_fname`) to runtime parameters (though, if possible, I'd prefer to keep them as hard-coded parameters to keep the UX simple.) 
 - In my Python prototype, memoization of determine_land_blocks sped up the iteration massively. If iteration performance becomes an issue in high resolution domains (e.g., in 1/12 deg), we may consider memoizing the Fortran version of determine_land_blocks subroutine as well.

### Testing

Ongoing. No answer changes and no issues so far.