@marshallward 2023-11-22
Thanks for this @alperaltuntas, this looks potentially very useful in contexts where one is not constraints to a particular layout.

I added some specific comments in the code, and have some general thoughts below.

* I am guessing that this ignores `LAYOUT` and `IO_LAYOUT` if the `AUTO_MASKTABLE` is turned on, meaning that it would grab `npes` from the launcher (`mpirun`, `srun`, etc.)  What happens if `LAYOUT` is set and there's a conflict?  Should there be an error?  I would discourage a `WARNING`, since they often get ignored and lost in the stdout bloat.

  Note that `LAYOUT` can often be chosen for architectural reasons and might interfere with the ability to use the auto masktable.

* It appears to me that `gen_auto_mask_table` would be run on every re-submission, even though the result should not change.  This could potentially be stored in the output, right?  (Maybe this is what you meant by memoization.)  If I am wrong about this and it is stored and reused, then please disregard.

* On that note, could the static `MOM_mask_table` and generated `MOM_auto_mask_table` be merged into some common file (or "format")?  Seems like the information ought to be similar.

* I don't think there's any problem with keeping these subroutines in the `MOM_domains` module.  Moving to `MOM_domains_infra` would seem incorrect, since there are no framework-dependent operations here.  And if there's a way to move them without circular dependencies, we can do it in the future.

* No thoughts yet on converting `ibuf`, `r_extreme`, etc into parameters, expectially if you feel the current values are already well-tuned.  If it does ever happen, I'd probably recommend storing them in a parameter block to avoid bloat to `MOM_parameter_doc.all`.

* The 23% speedup: Does it refer to fewer CPU cycles?  Or an actual reduction of runtime?  In other words, if you had defined the layout and preprocessed this land mask, would you have gotten the same result?

* See the inline comments on dimensionality, I have a feeling you will need to address this at some point.

Unfortunately I am short on time right now, but these are my thoughts for the moment.  Feel free to act on them as you wish :smile:.

Also, this has absolutely no bearing on the PR, but I like using `ib/ie` (as `ibegin/end`) in place of `is/ie` (as `istart/end`).  Using `is` as a variable drives me crazy!