@alperaltuntas 2023-11-22
Thanks, @marshallward!

I couldn't locate your inline comments, but here are my quick responses to your bulletpoints above.

>  am guessing that this ignores LAYOUT and IO_LAYOUT if the AUTO_MASKTABLE is turned on, meaning that it would grab npes from the launcher (mpirun, srun, etc.) What happens if LAYOUT is set and there's a conflict? Should there be an error? I would discourage a WARNING, since they often get ignored and lost in the stdout bloat.

Right, `LAYOUT` and `IO_LAYOUT` are ignored when `AUTO_MASKTABLE` is on, in which case npes is grabbed via `MOM_coms_infra:: num_PEs`. And, `LAYOUT` is auto-determined at runtime to maximize the number of eliminated land blocks. Similarly, `IO_LAYOUT` is determined at runtime if `AUTO_IO_LAYOUT_FAC` parameter is specified. Otherwise it's set to `1,1`. It's a good idea to throw an error when there is a discrepancy. Will do.

> It appears to me `that gen_auto_mask_table` would be run on every re-submission, even though the result should not change. This could potentially be stored in the output, right? (Maybe this is what you meant by memoization.) If I am wrong about this and it is stored and reused, then please disregard.

Right, `gen_auto_mask_table` is run on every re-submission if `AUTO_MASKTABLE` remains True. I thought it would be fine to do so because it only takes around ~0.1 sec of runtime, and doing so would prevent the usage of outdated mask tables when the user changes things like the PE count, the topography, or minimum depth. 

> On that note, could the static MOM_mask_table and generated MOM_auto_mask_table be merged into some common file (or "format")? Seems like the information ought to be similar.

Indeed, `MOM_auto_mask_table` has the same format as `MOM_mask_table`. And, from FMS's point of view, there is no difference between the two.

> I don't think there's any problem with keeping these subroutines in the MOM_domains module. Moving to MOM_domains_infra would seem incorrect, since there are no framework-dependent operations here. And if there's a way to move them without circular dependencies, we can do it in the future.

Sounds good!

> The 23% speedup: Does it refer to fewer CPU cycles? Or an actual reduction of runtime?

It refers to the reduction of runtime when compared to a run with no (static or automated) masking. So, it just shows the impact of masking on the wallclock runtime.

>  In other words, if you had defined the layout and preprocessed this land mask, would you have gotten the same result?

Correct.

>  See the inline comments on dimensionality

Unfortunately, I can't see your inline comments for some reason.

> Using is as a variable drives me crazy!

I agree!