@marshallward 2023-11-22
Sorry about that, I thought my overall comment was in the report.  I've just submitted it.

> Right, gen_auto_mask_table is run on every re-submission if AUTO_MASKTABLE remains True. I thought it would be fine to do so because it only takes around ~0.1 sec of runtime, and doing so would prevent the usage of outdated mask tables when the user changes things like the PE count, the topography, or minimum depth.

If the runtime is this small, then perhaps it's not too important.  (I believe I misinterpreted a comment somewhere.)

It might be meaningful to not enable the automasking if a `MOM_mask_table` is present and enabled in `MOM_input`.  I don't have strong feelings about this, but maybe others do.

Aside from the inline comments (which have now been submitted), I think this looks good.